Make Slf4j plugin default logger provider
