+++ src/main/java/io/vlingo/actors/Logger.java
@@ -8,15 +8,15 @@
 package io.vlingo.actors;
 
 import io.vlingo.actors.plugin.logging.noop.NoOpLogger;
-import io.vlingo.actors.plugin.logging.slf4j.Slf4jLogger;
+import io.vlingo.actors.plugin.logging.slf4j.Slf4jLoggerPlugin;
 
 public interface Logger {
-  public static Logger noOpLogger() {
+  static Logger noOpLogger() {
     return new NoOpLogger();
   }
 
-  public static Logger basicLogger() {
-    return Slf4jLogger.basicInstance();
+  static Logger basicLogger() {
+    return Slf4jLoggerPlugin.basicInstance();
   }
   
   String name();
